+++ nix/shell.nix
@@ -106,7 +106,7 @@
         pkgs = [
           {
             name = "cardano-cli";
-            help = "CLI v9.1.0 that is used in partner-chains dependency stack";
+            help = "CLI v9.1.1 that is used in partner-chains dependency stack";
             # This command has some eval because of IFD
             command = "${self'.packages.cardano-cli}/bin/cardano-cli $@";
           }